test: panic on PC 0 or opcode 0x0 & rustfmt
panic if PC is reset to 0 or the opcode is 0x0. This may change in the
future but 8080PRE resets PC to 0 to report errors.

Also rustfmt changes & clean up.

+++ src/memory.rs
@@ -44,8 +44,6 @@ impl Memory {
     }
     pub fn read_word(&mut self, addr: u16) -> u16 {
         return (self.memory[addr as usize] | (self.memory[(addr as usize + 1) << 8])) as u16;
-
-        // return self.read_byte(addr as u8) | (self.read_byte(addr  as u8 + 1) << 8)
     }
     pub fn write_memory(&mut self, addr: u16) {
         // (self.read_byte(addr + 2) as u16) << 8 | self.read_byte(addr + 1) as u16
@@ -71,8 +69,8 @@ impl Memory {
     }
 
     // Useful to read values out of memory to assign to 8 bit registers
-    pub fn read(&mut self, addr: usize) -> u8 {
-        self.memory[addr] as u8
+    pub fn read(&mut self, addr: u16) -> u8 {
+        self.memory[addr as usize] as u8
     }
     pub fn write(&mut self, addr: usize, val: u8) {
         self.memory[addr] = val;
